requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.5.2
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.4.0
carvekit-colab==4.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
datasets==2.4.0
decorator==5.1.1
diffusers==0.12.1
dill==0.3.5.1
dl-ext==1.3.4
dnspython==2.6.1
easydict==1.9
einops==0.3.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastcore==1.5.54
ffmpy==0.3.2
filelock==3.14.0
fire==0.4.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gradio==3.21.0
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.1
hydra-core==1.1.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib_metadata==8.1.0
importlib_resources==6.4.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
loguru==0.7.2
lovely-numpy==0.2.13
lovely-tensors==0.1.16
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.4.0
mmcv-full==1.7.2
mmdet==2.3.0
mmengine==0.10.4
model-index==0.1.11
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.13
namex==0.0.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.23.3
omegaconf==2.1.1
open_clip_torch==2.26.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
orjson==3.10.6
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.13.1
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pudb==2019.2
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.4.2
pytz==2023.4
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.28.2
responses==0.18.0
rich==13.4.2
rpds-py==0.19.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.9.1
shapely==2.0.5
shellingham==1.5.4
simple-lama-inpainting==0.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
super-image==0.1.7
tabulate==0.9.0
taming-transformers @ file:///mnt/sravanth/Adhoc/zero123/zero123/taming-transformers
taming-transformers-rom1504==0.0.6
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminaltables==3.1.10
test_tube==0.7.5
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==1.13.1+cu117
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.14.1+cu117
tornado==6.4.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.43.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==1.26.19
urwid==2.6.15
uvicorn==0.30.3
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.5
websockets==12.0
Werkzeug==3.0.3
wldhx.yadisk-direct==0.0.6
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
